requirements.txt
blinker==1.4
click==6.7
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.10.3
MarkupSafe==1.0
psycopg2-binary==2.7.5
SQLAlchemy>=1.3.0
werkzeug>=0.15.3
requests==2.22.0
xmltodict==0.12.0
python-dateutil==2.8.1
